+++ src/main/scala/li/cil/oc/util/InternetFilteringRule.scala
@@ -61,9 +61,6 @@ class InternetFilteringRule(val ruleString: String) {
                   val ipAddress = InetAddresses.forString(ipStringParts(0))
                   predicates += ((inetAddress: InetAddress, _: String) => ipAddress.equals(inetAddress))
                 }
-                predicates += ((inetAddress: InetAddress, _: String) => {
-                  inetAddress.isAnyLocalAddress || inetAddress.isLoopbackAddress || inetAddress.isLinkLocalAddress || inetAddress.isSiteLocalAddress
-                })
               case "all" =>
             }
           })
@@ -114,6 +111,7 @@ object InternetFilteringRule {
     "::1/128",
     "::ffff:0:0/96",
     "::/96",
+    "64:ff9b::/96", // NAT64 well-known prefix (RFC 6052)
     "100::/64",
     "2001:10::/28",
     "2001:db8::/32",