+++ changelog.md
@@ -4,6 +4,8 @@
 * Fixed Blood Altars (Blood Magic) returning the incorrect value for getSacrificeMultiplier. (hinyb)
 * Fixed client leak in Sound.
 * Fixed clipboard length not being validated on the server side.
+* Fixed network filtering not protecting against NAT64 accesses by default.
+* Fixed non-local addresses not being specifiable by the ip: filtering rule.
 * Fixed server leaks in Player, SideTracker and StaticSimpleEnvironment. (Alexdoru)
 * glGetError() is no longer called if logging GL errors is disabled. (mitchej123)
 * Improved Polish translation.